Break-up checks/statistics, more statistics, tests and cleanup
