+++ libcovebods/config.py
@@ -25,12 +25,42 @@ LIB_COVE_BODS_CONFIG_DEFAULT = {
     },
     # In some cases we default to the latest schema version, so we need to know what the latest version is.
     # It should be a key that is in the 'schema_versions' data.
+    # Note: Schema changed substantially from 0.3 to 0.4 (record based) so in cases where the schema
+    # version is not specified will try to determine whether is pre or post that change, and default to
+    # either the latest version (if after change) or a pre-change schema if before change
     "schema_latest_version": "0.4",
     # Or latest version before change to records (in 0.4)
     "schema_latest_nonrecord_version": "0.3",
     # These default values are very wide on purpose. It is left to apps using this to tighten them up.
-    "bods_additional_checks_person_birthdate_min_year": 1,
+    "bods_additional_checks_person_birthdate_min_year": 1800,
     "bods_additional_checks_person_birthdate_max_year": datetime.datetime.now().year,
+    # Other widely used (but "non-standard") country codes that should be allowed
+    # See: https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3 (User-assigned code elements section)
+    "bods_additional_checks_other_country_codes": (
+        "BAH",
+        "D",
+        "EUE",
+        "GBD",
+        "GBN",
+        "GBO",
+        "GBP",
+        "GBS",
+        "UNA",
+        "UNK",
+        "UNO",
+        "XBA",
+        "XIM",
+        "XCC",
+        "XCO",
+        "XEC",
+        "XPO",
+        "XOM",
+        "XXA",
+        "XXB",
+        "XXC",
+        "XXX",
+        "ZIM",
+    ),
 }
