+++ libcovebods/run_tasks.py
@@ -1,86 +1,98 @@
 import libcovebods.data_reader
-import libcovebods.tasks.checks
 import libcovebods.tasks.peps
-import libcovebods.tasks.statistics
+from libcovebods.tasks.checks import (
+    legacy_checks,
+    pre_record_checks,
+    record_based_checks,
+)
+from libcovebods.tasks.statistics import pre_record_statistics, record_based_statistics
 from libcovebods.utils import get_statement_type
 
 TASK_CLASSES = [
-    libcovebods.tasks.checks.LegacyChecks,
-    libcovebods.tasks.checks.LegacyChecksNeedingHistory,
-    libcovebods.tasks.checks.CheckHasPublicListing,
-    libcovebods.tasks.checks.CheckHasPublicListingRecord,
-    libcovebods.tasks.checks.CheckEntityTypeAndEntitySubtypeAlign,
-    libcovebods.tasks.checks.CheckEntitySecurityListingsMICSCodes,
-    libcovebods.tasks.checks.CheckEntitySecurityListingsMICSCodesRecord,
-    libcovebods.tasks.checks.CheckSourceRetrievedAtFutureDate,
-    libcovebods.tasks.checks.CheckStatementDateFutureDate,
-    libcovebods.tasks.checks.CheckAnnotationCreationDateFutureDate,
-    libcovebods.tasks.checks.CheckStatementPublicationDateFutureDate,
-    libcovebods.tasks.checks.CheckStatementPersonDateOfDeathSane,
-    libcovebods.tasks.checks.CheckStatementEntityFoundationDissolutionDates,
-    libcovebods.tasks.checks.CheckStatementPersonBirthDateSensible,
-    libcovebods.tasks.checks.CheckStatementRelationshipInterestsStartEndDates,
-    libcovebods.tasks.checks.CheckStatementRelationshipInterestsShareValues,
-    libcovebods.tasks.checks.CheckStatementDeclarationSubject,
-    libcovebods.tasks.checks.CheckStatementIsComponent,
-    libcovebods.tasks.checks.CheckStatementDuplicateStatementId,
-    libcovebods.tasks.checks.CheckStatementSeries,
-    libcovebods.tasks.checks.CheckStatementRelationshipParties,
-    libcovebods.tasks.checks.CheckAnnotationStatementPointerTarget,
-    libcovebods.tasks.checks.CheckStatementRelationshipInterests,
-    libcovebods.tasks.checks.CheckStatementSerialisation,
-    libcovebods.tasks.checks.CheckStatementPersonIdentifiersHaveCorrectScheme,
-    libcovebods.tasks.checks.CheckStatementEntityIdentifiersHaveKnownScheme,
-    libcovebods.tasks.statistics.StatisticsCountEntityStatements,
-    libcovebods.tasks.statistics.StatisticsCountEntityRecordStatements,
-    libcovebods.tasks.statistics.StatisticsCountPersonStatements,
-    libcovebods.tasks.statistics.StatisticsCountPersonRecordStatements,
-    libcovebods.tasks.statistics.StatisticsCountOwnershipOrControlStatements,
-    libcovebods.tasks.statistics.StatisticsCountOwnershipOrControlRecordStatements,
-    libcovebods.tasks.statistics.StatisticsCurrentOwnershipOrControlStatementsAndReplacesStatementsMissing,
-    libcovebods.tasks.statistics.StatisticAddress,
-    libcovebods.tasks.statistics.StatisticOwnershipOrControlInterestDirectOrIndirect,
-    libcovebods.tasks.statistics.StatisticOwnershipOrControlWithAtLeastOneInterestBeneficial,
-    libcovebods.tasks.statistics.StatisticDeclarationSubjects,
-    libcovebods.tasks.statistics.StatisticsStatementsRecordStatus,
+    legacy_checks.LegacyChecks,
+    legacy_checks.LegacyChecksNeedingHistory,
+    pre_record_checks.CheckHasPublicListing,
+    record_based_checks.CheckHasPublicListingRecord,
+    pre_record_checks.CheckEntityTypeAndEntitySubtypeAlign,
+    pre_record_checks.CheckEntitySecurityListingsMICSCodes,
+    record_based_checks.CheckEntitySecurityListingsMICSCodesRecord,
+    record_based_checks.CheckSourceRetrievedAtFutureDate,
+    record_based_checks.CheckStatementDateFutureDate,
+    record_based_checks.CheckAnnotationCreationDateFutureDate,
+    record_based_checks.CheckStatementPublicationDateFutureDate,
+    record_based_checks.CheckStatementPersonDateOfDeathSane,
+    record_based_checks.CheckStatementEntityFoundationDissolutionDates,
+    record_based_checks.CheckStatementPersonBirthDateSensible,
+    record_based_checks.CheckStatementRelationshipInterestsStartEndDates,
+    record_based_checks.CheckStatementRelationshipInterestsShareValues,
+    record_based_checks.CheckStatementDeclarationSubject,
+    record_based_checks.CheckStatementIsComponent,
+    record_based_checks.CheckStatementDuplicateStatementId,
+    record_based_checks.CheckStatementSeries,
+    record_based_checks.CheckStatementRelationshipParties,
+    record_based_checks.CheckAnnotationStatementPointerTarget,
+    record_based_checks.CheckStatementRelationshipInterests,
+    record_based_checks.CheckStatementSerialisation,
+    record_based_checks.CheckStatementPersonIdentifiersHaveCorrectScheme,
+    record_based_checks.CheckStatementEntityIdentifiersHaveKnownScheme,
+    pre_record_statistics.StatisticsCountEntityStatements,
+    record_based_statistics.StatisticsCountEntityRecordStatements,
+    pre_record_statistics.StatisticsCountPersonStatements,
+    record_based_statistics.StatisticsCountPersonRecordStatements,
+    pre_record_statistics.StatisticsCountOwnershipOrControlStatements,
+    record_based_statistics.StatisticsCountOwnershipOrControlRecordStatements,
+    pre_record_statistics.StatisticsCurrentOwnershipOrControlStatementsAndReplacesStatementsMissing,
+    pre_record_statistics.StatisticAddress,
+    record_based_statistics.StatisticRecordAddress,
+    pre_record_statistics.StatisticOwnershipOrControlInterestDirectOrIndirect,
+    record_based_statistics.StatisticOwnershipOrControlRecordInterestDirectOrIndirect,
+    pre_record_statistics.StatisticOwnershipOrControlWithAtLeastOneInterestBeneficial,
+    record_based_statistics.StatisticOwnershipOrControlRecordWithAtLeastOneInterestBeneficial,
+    record_based_statistics.StatisticDeclarationSubjects,
+    record_based_statistics.StatisticsStatementsRecordStatus,
     libcovebods.tasks.peps.PEPForSchema02Only,
     libcovebods.tasks.peps.PEPForSchema03AndAbove,
+    libcovebods.tasks.peps.PEPForSchema04AndAbove,
 ]
 
 TASK_CLASSES_IN_SAMPLE_MODE = [
-    libcovebods.tasks.checks.LegacyChecks,
-    libcovebods.tasks.checks.CheckHasPublicListing,
-    libcovebods.tasks.checks.CheckHasPublicListingRecord,
-    libcovebods.tasks.checks.CheckEntityTypeAndEntitySubtypeAlign,
-    libcovebods.tasks.checks.CheckEntitySecurityListingsMICSCodes,
-    libcovebods.tasks.checks.CheckEntitySecurityListingsMICSCodesRecord,
-    libcovebods.tasks.checks.CheckSourceRetrievedAtFutureDate,
-    libcovebods.tasks.checks.CheckStatementDateFutureDate,
-    libcovebods.tasks.checks.CheckAnnotationCreationDateFutureDate,
-    libcovebods.tasks.checks.CheckStatementPublicationDateFutureDate,
-    libcovebods.tasks.checks.CheckStatementPersonDateOfDeathSane,
-    libcovebods.tasks.checks.CheckStatementEntityFoundationDissolutionDates,
-    libcovebods.tasks.checks.CheckStatementPersonBirthDateSensible,
-    libcovebods.tasks.checks.CheckStatementRelationshipInterestsStartEndDates,
-    libcovebods.tasks.checks.CheckStatementRelationshipInterestsShareValues,
-    libcovebods.tasks.checks.CheckStatementDuplicateStatementId,
-    libcovebods.tasks.checks.CheckAnnotationStatementPointerTarget,
-    libcovebods.tasks.checks.CheckStatementRelationshipInterests,
-    libcovebods.tasks.checks.CheckStatementPersonIdentifiersHaveCorrectScheme,
-    libcovebods.tasks.checks.CheckStatementEntityIdentifiersHaveKnownScheme,
-    libcovebods.tasks.statistics.StatisticsCountEntityStatements,
-    libcovebods.tasks.statistics.StatisticsCountEntityRecordStatements,
-    libcovebods.tasks.statistics.StatisticsCountPersonStatements,
-    libcovebods.tasks.statistics.StatisticsCountPersonRecordStatements,
-    libcovebods.tasks.statistics.StatisticsCountOwnershipOrControlStatements,
-    libcovebods.tasks.statistics.StatisticsCountOwnershipOrControlRecordStatements,
-    libcovebods.tasks.statistics.StatisticAddress,
-    libcovebods.tasks.statistics.StatisticOwnershipOrControlInterestDirectOrIndirect,
-    libcovebods.tasks.statistics.StatisticOwnershipOrControlWithAtLeastOneInterestBeneficial,
-    libcovebods.tasks.statistics.StatisticDeclarationSubjects,
-    libcovebods.tasks.statistics.StatisticsStatementsRecordStatus,
+    legacy_checks.LegacyChecks,
+    pre_record_checks.CheckHasPublicListing,
+    record_based_checks.CheckHasPublicListingRecord,
+    pre_record_checks.CheckEntityTypeAndEntitySubtypeAlign,
+    pre_record_checks.CheckEntitySecurityListingsMICSCodes,
+    record_based_checks.CheckEntitySecurityListingsMICSCodesRecord,
+    record_based_checks.CheckSourceRetrievedAtFutureDate,
+    record_based_checks.CheckStatementDateFutureDate,
+    record_based_checks.CheckAnnotationCreationDateFutureDate,
+    record_based_checks.CheckStatementPublicationDateFutureDate,
+    record_based_checks.CheckStatementPersonDateOfDeathSane,
+    record_based_checks.CheckStatementEntityFoundationDissolutionDates,
+    record_based_checks.CheckStatementPersonBirthDateSensible,
+    record_based_checks.CheckStatementRelationshipInterestsStartEndDates,
+    record_based_checks.CheckStatementRelationshipInterestsShareValues,
+    record_based_checks.CheckStatementDuplicateStatementId,
+    record_based_checks.CheckAnnotationStatementPointerTarget,
+    record_based_checks.CheckStatementRelationshipInterests,
+    record_based_checks.CheckStatementPersonIdentifiersHaveCorrectScheme,
+    record_based_checks.CheckStatementEntityIdentifiersHaveKnownScheme,
+    pre_record_statistics.StatisticsCountEntityStatements,
+    record_based_statistics.StatisticsCountEntityRecordStatements,
+    pre_record_statistics.StatisticsCountPersonStatements,
+    record_based_statistics.StatisticsCountPersonRecordStatements,
+    pre_record_statistics.StatisticsCountOwnershipOrControlStatements,
+    record_based_statistics.StatisticsCountOwnershipOrControlRecordStatements,
+    pre_record_statistics.StatisticAddress,
+    record_based_statistics.StatisticRecordAddress,
+    pre_record_statistics.StatisticOwnershipOrControlInterestDirectOrIndirect,
+    record_based_statistics.StatisticOwnershipOrControlRecordInterestDirectOrIndirect,
+    pre_record_statistics.StatisticOwnershipOrControlWithAtLeastOneInterestBeneficial,
+    record_based_statistics.StatisticOwnershipOrControlRecordWithAtLeastOneInterestBeneficial,
+    record_based_statistics.StatisticDeclarationSubjects,
+    record_based_statistics.StatisticsStatementsRecordStatus,
     libcovebods.tasks.peps.PEPForSchema02Only,
     libcovebods.tasks.peps.PEPForSchema03AndAbove,
+    libcovebods.tasks.peps.PEPForSchema04AndAbove,
 ]
