+++ tests/test_stat_count_addresses_0_4.py
@@ -1,0 +1,29 @@
+import os
+import tempfile
+
+from tests.api import bods_json_output
+
+
+def test_stat_count_addresses_1():
+
+    cove_temp_folder = tempfile.mkdtemp(
+        prefix="lib-cove-bods-tests-", dir=tempfile.gettempdir()
+    )
+    json_filename = os.path.join(
+        os.path.dirname(os.path.realpath(__file__)),
+        "fixtures",
+        "0.4",
+        "sample_300_statements.json",
+    )
+
+    results = bods_json_output(cove_temp_folder, json_filename)
+
+    assert results["schema_version"] == "0.4"
+
+    assert results["statistics"]["count_addresses"] == 100
+    assert results["statistics"]["count_addresses_with_postcode"] == 100
+    assert results["statistics"]["count_addresses_with_country"] == 100
+    assert (
+        results["statistics"]["count_addresses_with_postcode_duplicated_in_address"]
+        == 0
+    )
